update service package add LabelService.java and TaskService.java

+++ app/src/main/java/hexlet/code/service/TaskStatusService.java
@@ -3,22 +3,23 @@ package hexlet.code.service;
 import hexlet.code.dto.TaskStatusDTO.TaskStatusCreateDTO;
 import hexlet.code.dto.TaskStatusDTO.TaskStatusDTO;
 import hexlet.code.dto.TaskStatusDTO.TaskStatusUpdateDTO;
-import hexlet.code.exception.UserNotFoundException;
+import hexlet.code.exception.MethodNotAllowedException;
+import hexlet.code.exception.ResourceNotFoundException;
 import hexlet.code.mapper.TaskStatusMapper;
+import hexlet.code.repository.TaskRepository;
 import hexlet.code.repository.TaskStatusRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
+@AllArgsConstructor
 public class TaskStatusService {
 
-    @Autowired
     private TaskStatusRepository taskStatusRepository;
-
-    @Autowired
     private TaskStatusMapper taskStatusMapper;
+    private TaskRepository taskRepository;
 
     public List<TaskStatusDTO> getAllStatuses() {
         return taskStatusRepository.findAll().stream()
@@ -34,19 +35,24 @@ public class TaskStatusService {
 
     public TaskStatusDTO findById(Long statusId) {
         var taskStatus = taskStatusRepository.findById(statusId)
-                .orElseThrow(() -> new UserNotFoundException("Status with id: " + statusId + " not found."));
+                .orElseThrow(() -> new ResourceNotFoundException("Status with id: " + statusId + " not found."));
         return taskStatusMapper.map(taskStatus);
     }
 
     public TaskStatusDTO updateStatus(Long statusId, TaskStatusUpdateDTO data) {
         var taskStatus = taskStatusRepository.findById(statusId)
-                .orElseThrow(() -> new UserNotFoundException("Status with id: " + statusId + " not found."));
+                .orElseThrow(() -> new ResourceNotFoundException("Status with id: " + statusId + " not found."));
         taskStatusMapper.update(data, taskStatus);
         taskStatusRepository.save(taskStatus);
         return taskStatusMapper.map(taskStatus);
     }
 
     public void deleteStatus(Long statusId) {
+        var taskStatus = taskStatusRepository.findById(statusId);
+
+        if (taskStatus.isPresent() && taskRepository.findByTaskStatusName(taskStatus.get().getName()).isPresent()) {
+            throw new MethodNotAllowedException("TaskStatus still has task");
+        }
         taskStatusRepository.deleteById(statusId);
     }
 }
